All Edit Pane is now in ReactList instead of ActionNews component
