+++ src/webparts/actionnews/components/Actionnews.tsx
@@ -39,8 +39,6 @@ import  EarlyAccess from './HelpInfo/EarlyAccess';
 
 import ThisEditPane from './Panel/ThisEditPane';
 
-import { _saveEditPaneItem } from './Panel/editPaneFunctions';
-
 import { ActionQuickCommands } from './const_ActionCommands';
 
 import { ActionNewsViewDefs } from './const_ActionViewDefs';
@@ -48,10 +46,6 @@ import { ActionNewsViewDefs } from './const_ActionViewDefs';
 import { getNewActionQuickFields } from './const_ActionQuickFields';
 
 import { findParentElementPropLikeThis } from '../../../services/basicElements';
-
-import { msPerWk, msPerDay } from '../../../services/dateServices';
-
-import { getEmailFromLoginName, checkForLoginName, ensureUserHere, ensureTheseUsers } from '../../../services/userServices';
 
 import { getAppropriateViewFields, getAppropriateViewGroups, } from './ReactList/listFunctions';
 import { PrincipalType } from '@pnp/spfx-controls-react/lib/PeoplePicker';
@@ -226,15 +220,6 @@ let quickCommands : IQuickCommands = ActionQuickCommands;
 
   let quickFields : IQuickField[][] = getNewActionQuickFields( this.makeDefaultTextField('Title','Fetching Page Title - '), this.makeDefaultTextField('Comments',this.props.comments ) ) ;
 
-  let updateQuickFields = false;
-  if ( !quickCommands ) { }  //do nothing 
-  else if ( !quickCommands.quickFields ) { updateQuickFields = true ; }
-  else if ( quickCommands.quickFields.length === 0 ) { updateQuickFields = true ; }
-
-  if ( updateQuickFields === true ) {
-    quickCommands.quickFields = quickFields;
-  }
-
   this.state = {
 
         // 0 - Context
@@ -349,7 +334,6 @@ public componentDidUpdate(prevProps){
     let toggleTipsButton = <div style={{marginRight: "20px", background: 'white', opacity: '.7', borderRadius: '10px' }}>
     { createIconButton('Help','Toggle Tips',this.toggleTips.bind(this), null, tipsStyles ) } </div>;
 
-    let toggleNewItemPane = createIconButton('CirclePlus','Create Item',this._onShowPanelNewItem.bind(this), null, tipsStyles );
     //let toggleTipsButton = createIconButton('Help','Toggle Tips',this.toggleTips.bind(this), null, null );
 
     /***
@@ -372,42 +356,6 @@ public componentDidUpdate(prevProps){
       ></InfoPage>
     </div>;
 
-
-    let newPanel = this.state.showNewItem !== true ? null : 
-    <Panel
-          isOpen={this.state.showNewItem}
-          type={ this.state.panelWidth }
-          onDismiss={this._onClosePanelNewItem}
-          headerText={ 'Create New ActionNews Item' }
-          closeButtonAriaLabel="Close"
-          onRenderFooterContent={this._onRenderFooterContent}
-          onRenderHeader={ this.props.allowSplit ? this._onRenderHeader : null }
-          isLightDismiss={ true }
-          isFooterAtBottom={ true }
-      >
-
-        <ThisEditPane 
-            wpContext={ this.props.wpContext }
-            webAbsoluteUrl={ this.state.newsService.listWeb }
-            fields = { this.state.quickFields }
-            contextUserInfo = { this.state.newsService.contextUserInfo }
-            sourceUserInfo = { this.state.newsService.sourceUserInfo }
-            onChange = { this._editFieldUpdate.bind(this) }
-            _clearDateField = { this._clearDateField.bind(this) }
-            _addYouToField = { this._addUserToField.bind(this) }
-            _addWeekToDate = { this._addWeekToDate.bind(this) }
-            _updateDropdown = { this._updateDropdown.bind(this) }
-            _saveItem= { this._saveItem.bind(this) }
-            _cancelItem= { this._onClosePanelNewItem.bind(this) }
-            allowSplit= { this.state.allowSplit }
-            _getTitleValue = { null /*this.updatePageTitleInStateTest.bind(this)  null */  }
-            readOnlyMode = { false }
-            
-
-        ></ThisEditPane>
-
-    </Panel>;
-
     /***
      *    db      d888888b .d8888. d888888b      d888888b d888888b d88888b .88b  d88. .d8888. 
      *    88        `88'   88'  YP `~~88~~'        `88'   `~~88~~' 88'     88'YbdP`88 88'  YP 
@@ -421,7 +369,7 @@ public componentDidUpdate(prevProps){
  
     let actionNewsHeader = <div style={{ float: 'right' }}>
         <Stack horizontal={true} wrap={true} horizontalAlign={"end"} verticalAlign= {"center"} tokens={stackPageTokens}>{}
-          { toggleNewItemPane }
+
         </Stack>
       </div>;
 
@@ -430,32 +378,41 @@ public componentDidUpdate(prevProps){
 
     let currentViewGroups : IGrouping[] =  getAppropriateViewGroups( ActionNewsViewDefs , this.state.WebpartWidth );
 
-    let  actionNewsItems  = this.state.allItems.length === 0 ? <div>NO ITEMS FOUND</div> :
+    let  actionNewsItems  = 
       <ReactListItems
           parentListFieldTitles={ ActionNewsViewDefs.length > 0 ? null : null }
 
+          blueBar = { 'Action Items' }
           webURL = { this.state.newsService.listWeb }
           parentListURL = { this.state.newsService.listWeb + '/lists/' + this.state.newsService.listName }
           listName = { this.state.newsService.listName }
 
+          wpContext={ this.props.wpContext }
           contextUserInfo = { this.state.newsService.contextUserInfo }
           sourceUserInfo = { this.state.newsService.sourceUserInfo }
 
-          viewFields={ currentViewFields }
-          groupByFields={ currentViewGroups }
-          items={ this.state.allItems }
-          includeDetails= { true }
-          includeAttach= { false }
-          includeListLink = { true }
-          quickCommands={ this.state.quickCommands }
+          recentUsers = { this.state.recentUsers }
+          viewFields = { currentViewFields }
+          groupByFields = { currentViewGroups }
+          items = { this.state.allItems }
 
-          updateParentComponentEditPane= { this._onShowPanelEditItem.bind(this) }
-          showPanel= { false }
+          reloadAllItems = { this.getAllItemsCall.bind(this) }
+
+          includeDetails = { true }
+          includeAttach = { false }
+          includeListLink = { true }
+          quickCommands = { this.state.quickCommands }
+
+          quickFields = { this.state.quickFields }
+          staticFields = { this.state.staticFields }
 
           //Size courtesy of https://www.netwoven.com/2018/11/13/resizing-of-spfx-react-web-parts-in-different-scenarios/
           WebpartHeight={ this.props.WebpartElement.getBoundingClientRect().height }
           WebpartWidth={ this.props.WebpartElement.getBoundingClientRect().width - 50 }
       
+
+          allowSplit={ this.state.allowSplit }
+
       ></ReactListItems>;
 
 
@@ -531,8 +488,6 @@ public componentDidUpdate(prevProps){
             </div>
             <div>
             <Stack horizontal={false} wrap={true} horizontalAlign={"stretch"} tokens={stackPageTokens} className={ styles.actionButtons }>{/* Stack for Buttons and Webs */}
-
-                { newPanel } 
                 { actionNewsHeader }
                 { actionNewsItems }
             </Stack>
@@ -577,28 +532,18 @@ public componentDidUpdate(prevProps){
         console.log('addTheseItemsToState allItems: QTY: ', allItems.length );
     }
 
-    let quickCommands = this.state.quickCommands;
-    quickCommands.quickFields = quickFields;
-
     this.setState({
         allItems: allItems,
         newsService:  newsService,
         errMessage: errMessage,
         quickFields: quickFields,
-        quickCommands: quickCommands,
         recentUsers: recentUsers,
     });
 
     return true;
   }
 
-  private updateQuickFieldsInState( quickFields : IQuickField[][] , quickCommands : IQuickCommands, showNewItem: boolean ) {
 
-    quickCommands.quickFields = quickFields;
-
-    this.setState({ quickFields: quickFields, quickCommands: quickCommands, showNewItem });
-
-  }
   /***
    *         db    db d8888b. d8888b.  .d8b.  d888888b d88888b      .d8888. d888888b  .d8b.  d888888b d88888b 
    *         88    88 88  `8D 88  `8D d8' `8b `~~88~~' 88'          88'  YP `~~88~~' d8' `8b `~~88~~' 88'     
@@ -629,67 +574,6 @@ public componentDidUpdate(prevProps){
     this.getAllItemsCall();
   }
 
-  
-/***
- *    d8888b.  .d8b.  d8b   db d88888b db           d88888b  .d88b.   .d88b.  d888888b d88888b d8888b. 
- *    88  `8D d8' `8b 888o  88 88'     88           88'     .8P  Y8. .8P  Y8. `~~88~~' 88'     88  `8D 
- *    88oodD' 88ooo88 88V8o 88 88ooooo 88           88ooo   88    88 88    88    88    88ooooo 88oobY' 
- *    88~~~   88~~~88 88 V8o88 88~~~~~ 88           88~~~   88    88 88    88    88    88~~~~~ 88`8b   
- *    88      88   88 88  V888 88.     88booo.      88      `8b  d8' `8b  d8'    88    88.     88 `88. 
- *    88      YP   YP VP   V8P Y88888P Y88888P      YP       `Y88P'   `Y88P'     YP    Y88888P 88   YD 
- *                                                                                                     
- *                                                                                                     
- */
-
-      /**
-       * This was copied from codepen example code to render a footer with buttons:
-       * https://fabricweb.z5.web.core.windows.net/oufr/6.50.2/#/examples/panel
-       * 
-       */
-      private _onRenderFooterContent = (): JSX.Element => {
-        return null;
-
-        return (
-        <div>
-            <PrimaryButton onClick={this._onClosePanelNewItem} style={{ marginRight: '8px' }}>
-            Save
-            </PrimaryButton>
-            <DefaultButton onClick={this._onClosePanelNewItem}>Cancel</DefaultButton>
-        </div>
-        );
-    }
-
-    private _onRenderHeader = (): JSX.Element => {
-
-      let defStyles = { root: { width: 160, } };
-
-      let thisToggle = <Toggle label={ 'Split Notifications' } 
-          onText={ 'On' } 
-          offText={ 'Off' } 
-          onChange={ this._toggleSplit.bind(this) } 
-          checked={ this.state.allowSplit }
-          styles={ defStyles }
-      />;
-      
-      const stackTokens: IStackTokens = { childrenGap: 20 };
-
-      return (
-      <div>
-        <Stack horizontal={ true } horizontalAlign= { 'space-between' } tokens={stackTokens}>
-          <span style={{ marginLeft: 35, fontSize: 28, marginTop: 5 }}>Create Action News item(s)</span>
-          { thisToggle }
-        </Stack>
-
-      </div>
-      );
-  }
-
-  private _toggleSplit() {
-
-    this.setState({ allowSplit: this.state.allowSplit !== true ? true : false });
-
-  }
-
 /***
  *    .d8888. db   db  .d88b.  db   d8b   db      d8888b.  .d8b.  d8b   db d88888b db      
  *    88'  YP 88   88 .8P  Y8. 88   I8I   88      88  `8D d8' `8b 888o  88 88'     88      
@@ -701,37 +585,6 @@ public componentDidUpdate(prevProps){
  *                                                                                         
  */
 
-  private _onClosePanelNewItem = (item): void => {
-    
-      this.setState({ 
-          showNewItem: false,
-      });
-  }
-
-
-  public _onShowPanelEditItem = ( item: any ): void => {
-    let quickFields : IQuickField[][] = JSON.parse(JSON.stringify( this.state.quickCommands.quickFields ));
-
-    quickFields.map( fieldRow => {
-      fieldRow.map( thisFieldObject => {
-            thisFieldObject.value = item[0][thisFieldObject.name];
-      });
-    }) ;
-
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, true );
-  } //End toggleNewItem 
-
-
-  public _onShowPanelNewItem = ( item: any ): void => {
-  // public async _onShowPanelNewItem ( item: any ) {
-    //This sends back the correct pivot category which matches the category on the tile.
-    //sourceUserInfo
-
-    this.setState({
-      showNewItem: true,
-    });
-
-  } //End toggleNewItem  
 
   public toggleTips = (item: any): void => {
     //This sends back the correct pivot category which matches the category on the tile.
@@ -742,255 +595,5 @@ public componentDidUpdate(prevProps){
 
   } //End toggleTips  
 
-/***
- *    d8888b. d88888b  .d88b.  d8888b. db      d88888b      d8888b. d888888b  .o88b. db   dD d88888b d8888b. 
- *    88  `8D 88'     .8P  Y8. 88  `8D 88      88'          88  `8D   `88'   d8P  Y8 88 ,8P' 88'     88  `8D 
- *    88oodD' 88ooooo 88    88 88oodD' 88      88ooooo      88oodD'    88    8P      88,8P   88ooooo 88oobY' 
- *    88~~~   88~~~~~ 88    88 88~~~   88      88~~~~~      88~~~      88    8b      88`8b   88~~~~~ 88`8b   
- *    88      88.     `8b  d8' 88      88booo. 88.          88        .88.   Y8b  d8 88 `88. 88.     88 `88. 
- *    88      Y88888P  `Y88P'  88      Y88888P Y88888P      88      Y888888P  `Y88P' YP   YD Y88888P 88   YD 
- *                                                                                                           
- *                                                                                                           
- */
-
-  private _addUserToField = (prop: string, valueX: any ): void => {
-    let e: any = event;
-    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
-    thisID = thisID.replace('EditFieldID','');
-
-    let quickFields = this.state.quickFields;
-
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.name === thisID ) { 
-          if (field.type.toLowerCase().indexOf('user') < 0 ) {
-            alert('Error in _addUserToField:  Trying to add user to non-user field!');
-          } else {
-            let value = field.value;
-            if ( value == null ) { value = []; }
-            value.push( this.state.newsService.contextUserInfo );
-            field.value = value;
-            this.updateRecentUsers( field.value, this.state.recentUsers, this.state.newsService.listWeb );
-          }
-        }
-      });
-    });
-    //Then update the quickFields
-
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, this.state.showNewItem );
-
-  }
-
-
-  /***
- *    d8888b.  .d8b.  d888888b d88888b      d8888b. d888888b  .o88b. db   dD d88888b d8888b. 
- *    88  `8D d8' `8b `~~88~~' 88'          88  `8D   `88'   d8P  Y8 88 ,8P' 88'     88  `8D 
- *    88   88 88ooo88    88    88ooooo      88oodD'    88    8P      88,8P   88ooooo 88oobY' 
- *    88   88 88~~~88    88    88~~~~~      88~~~      88    8b      88`8b   88~~~~~ 88`8b   
- *    88  .8D 88   88    88    88.          88        .88.   Y8b  d8 88 `88. 88.     88 `88. 
- *    Y8888D' YP   YP    YP    Y88888P      88      Y888888P  `Y88P' YP   YD Y88888P 88   YD 
- *                                                                                           
- *                                                                                           
- */
-
-  private _addWeekToDate = (prop: string, value: any ): void => {
-
-    let e: any = event;
-    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
-    thisID = thisID.replace('EditFieldID','');
-
-    let quickFields = this.state.quickFields;
-
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.name === thisID ) { 
-
-          //Based on https://www.sitepoint.com/community/t/how-do-i-add-one-week-to-a-date/47817/2
-          let start = field.value ? field.value: new Date();
-
-          field.value = new Date( start.getTime() + msPerDay * value );
-        }
-      });
-    });
-    //Then update the quickFields
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, this.state.showNewItem );
-
-  }
-
-  private _clearDateField = (prop: string, value: any ): void => {
-
-    let e: any = event;
-    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
-    thisID = thisID.replace('EditFieldID','');
-
-    let quickFields = this.state.quickFields;
-
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.name === thisID ) { field.value = null ;}
-      });
-    });
-    //Then update the quickFields
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, this.state.showNewItem );
-
-  }
-
-  /***
- *    d88888b d8888b. d888888b d888888b      db    db d8888b. d8888b.  .d8b.  d888888b d88888b 
- *    88'     88  `8D   `88'   `~~88~~'      88    88 88  `8D 88  `8D d8' `8b `~~88~~' 88'     
- *    88ooooo 88   88    88       88         88    88 88oodD' 88   88 88ooo88    88    88ooooo 
- *    88~~~~~ 88   88    88       88         88    88 88~~~   88   88 88~~~88    88    88~~~~~ 
- *    88.     88  .8D   .88.      88         88b  d88 88      88  .8D 88   88    88    88.     
- *    Y88888P Y8888D' Y888888P    YP         ~Y8888P' 88      Y8888D' YP   YP    YP    Y88888P 
- *                                                                                             
- *                                                                                             
- */
-
-  private _editFieldUpdate = ( prop: string, value: any ): void => {
-
-    let e: any = event;
-
-    let quickFields = this.state.quickFields;
-    
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.name === prop ) { 
-          field.value = value ;
-          console.log('found this item to update: ' , prop, value );
-
-          if ( field.type.toLowerCase().indexOf('user') === 0  ) {
-            this.updateRecentUsers( field.value, this.state.recentUsers, this.state.newsService.listWeb );
-
-          } else if ( field.type.toLowerCase().indexOf('user') > 0  ) { //covers multiUser and splitUser
-            this.updateRecentUsers( field.value, this.state.recentUsers, this.state.newsService.listWeb );
-
-          }
-        }
-      });
-    });
-    //Then update the quickFields
-
-    // console.log('HERE IS Current QuickFields:', quickFields );
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, this.state.showNewItem );
-
-  }
-
-  private async updateRecentUsers( theseUsers: IUser[], checkTheseUsers: IUser[] , webUrl: string ) {
-    let recentUsers = await ensureTheseUsers( theseUsers, checkTheseUsers, webUrl );
-    this.setState({
-      recentUsers: recentUsers,
-    });
-  }
-
-  private _updateDropdown = (prop: React.FormEvent<HTMLDivElement>, e , pickedOption ): void => {
-
-    let quickFields = this.state.quickFields;
-    let thisProp : any = prop;
-
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.name === thisProp ) { 
-          field.value = pickedOption.text ;
-          console.log('found this item to update: ' , thisProp, pickedOption.text );
-        }
-      });
-    });
-    //Then update the quickFields
-
-    // console.log('HERE IS Current QuickFields:', quickFields );
-    this.updateQuickFieldsInState( quickFields , this.state.quickCommands, this.state.showNewItem );
-
-  }
-
-/***
- *    .d8888.  .d8b.  db    db d88888b      d888888b d888888b d88888b .88b  d88. 
- *    88'  YP d8' `8b 88    88 88'            `88'   `~~88~~' 88'     88'YbdP`88 
- *    `8bo.   88ooo88 Y8    8P 88ooooo         88       88    88ooooo 88  88  88 
- *      `Y8b. 88~~~88 `8b  d8' 88~~~~~         88       88    88~~~~~ 88  88  88 
- *    db   8D 88   88  `8bd8'  88.            .88.      88    88.     88  88  88 
- *    `8888Y' YP   YP    YP    Y88888P      Y888888P    YP    Y88888P YP  YP  YP 
- *                                                                               
- *                                                                               
- */
-
-  private async _saveItem ( ) {
-
-    let splitCount = 1;
-    let splitUsers = [];
-    let splitField = '';
-    //Get array of split users
-    let quickFields = this.state.quickFields;
-    
-    //Search through each row and field for name:
-    quickFields.map( fieldRow => {
-      fieldRow.map ( field => {
-        if ( field.type.toLowerCase().indexOf('split') > -1 ) { 
-          splitUsers = JSON.parse( JSON.stringify( field.value ));
-          splitField = field.name;
-          splitCount = field.value ? field.value.length : 0;
-        }
-      });
-    });
-
-    let results : any = null;
-    if ( splitCount === 1 ) {
-      let recentUsers = JSON.parse(JSON.stringify( this.state.recentUsers )); // Needed to prevent it from getting over-written in this function somewhere
-      results = await _saveEditPaneItem( this.state.newsService.listWeb, this.state.newsService.listName, this.state.quickFields, this.state.staticFields, recentUsers );
-  
-    } else {
-
-      //Save each item individually - unless allowSplit !== true, then just set to first item in array
-      if ( splitCount > 1 && this.state.allowSplit !== true ) { splitCount = 1; }
-      for (let i = 0; i < splitCount; i++) {
-
-        quickFields.map( fieldRow => {
-          fieldRow.map ( field => {
-            if ( field.name === splitField ) { 
-              field.value = [ splitUsers[i] ];
-            }
-          });
-        });
-
-        let recentUsers = JSON.parse(JSON.stringify( this.state.recentUsers )); // Needed to prevent it from getting over-written in this function somewhere
-        results = await _saveEditPaneItem( this.state.newsService.listWeb, this.state.newsService.listName, quickFields, this.state.staticFields, recentUsers );
-
-      }
-
-    }
-
-    let passed = results && results.data ? true : false;
-
-    if ( passed !== true ) {
-      //The save did not happend
-      console.log('was NOT ABLE TO SAVE ITEM');
-      
-      //Put back original splitUsers array - NOT needed if I clear this field after save.
-      quickFields.map( fieldRow => {
-        fieldRow.map ( field => {
-          if ( field.name === splitField ) { 
-            field.value = splitUsers;
-          }
-        });
-      });
-
-     alert('Your item was not saved...\nPossibly because you have a splitPersonField type with no value.');
-
-    } else {
-      alert('Your Action News item was just saved!');
-      quickFields = getNewActionQuickFields( this.makeDefaultTextField('Title', this.state.newsService.pageTitle) , this.makeDefaultTextField('Comments', this.props.comments)  ) ;
-      
-      this.updateQuickFieldsInState( quickFields , this.state.quickCommands, false );
-
-      this.getAllItemsCall();
-
-    }
-    return null;
-
-  }
 
 }
