+++ src/webparts/actionnews/components/IReUsableInterfaces.ts
@@ -202,7 +202,7 @@ export interface IQuickCommands {
     listName?: string;
     styleRow?: any; //Valid react JSON object for style
     showWhenEvalTrue?: string;  //This is on all commands or individual buttons.  if eval resolves to true, button or commands are visible, else they are false.
-    quickFields: IQuickField[][];
+
 }
 
 export interface IQuickField extends Partial<IFieldDef> {
