+++ src/webparts/actionnews/components/ReactList/reactListView.tsx
@@ -4,6 +4,11 @@ import { Icon  } from 'office-ui-fabric-react/lib/Icon';
 
 import { Web, IList, IItem, } from "@pnp/sp/presets/all";
 import { Link, ILinkProps } from 'office-ui-fabric-react';
+import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
+
+import { WebPartContext } from '@microsoft/sp-webpart-base';
+
+import { PageContext } from '@microsoft/sp-page-context';
 
 import { IMyProgress, IQuickButton, IQuickCommands, IUser, IQuickField } from '../IReUsableInterfaces';
 
@@ -34,9 +39,21 @@ import { Label } 			from 'office-ui-fabric-react/lib/Label';
 
 import { updateReactListItem } from './listFunctions';
 
+import { createIconButton , defCommandIconStyles} from "../createButtons/IconButton";
+
 import { createPanelButtons, ButtonIdDelim } from '../Panel/panelFunctions';
 
+import { getEmailFromLoginName, checkForLoginName, ensureUserHere, ensureTheseUsers } from '../../../../services/userServices';
+
+import { msPerWk, msPerDay } from '../../../../services/dateServices';
+
+import ThisEditPane from '../Panel/ThisEditPane';
+
+import { _saveEditPaneItem } from '../Panel/editPaneFunctions';
+
 import { IContentsToggles, makeToggles } from '../fields/toggleFieldBuilder';
+
+import { getNewActionQuickFields } from '../const_ActionQuickFields';
 
 import styles from '../Contents/listView.module.scss';
 import stylesInfo from '../HelpInfo/InfoPane.module.scss';
@@ -50,6 +67,8 @@ export interface IReactListItemsProps {
     maxChars?: number;
     items: IActionItem[];
 
+    reloadAllItems: any;
+
     WebpartHeight:  number;    //Size courtesy of https://www.netwoven.com/2018/11/13/resizing-of-spfx-react-web-parts-in-different-scenarios/
     WebpartWidth:   number;    //Size courtesy of https://www.netwoven.com/2018/11/13/resizing-of-spfx-react-web-parts-in-different-scenarios/
 
@@ -57,10 +76,11 @@ export interface IReactListItemsProps {
     listName: string; //Used for attachments
     parentListURL: string;
 
+    wpContext: WebPartContext;
     contextUserInfo: IUser;  //For site you are on ( aka current page context )
     sourceUserInfo: IUser;   //For site where the list is stored
 
-    blueBar?: any;
+    blueBar: any;
 
     showIDs?: boolean;
     showDesc?: boolean;
@@ -76,9 +96,12 @@ export interface IReactListItemsProps {
     highlightedFields?: string[];
 
     quickCommands?: IQuickCommands;
+    quickFields: IQuickField[][];
+    staticFields: IQuickField[][];
 
-    updateParentComponentEditPane?: any;
-    showPanel: boolean;
+    allowSplit: boolean;
+
+    recentUsers: IUser[];
 
 }
 
@@ -87,8 +110,12 @@ export interface IReactListItemsState {
   parentListFieldTitles: any;
   viewFields: IViewField[];
   groupByFields?:  IGrouping[];
-  
+  quickFields: IQuickField[][];
+
   showPanel: boolean;
+  showNewPanel: boolean;
+  showEditPanel: boolean;
+  panelMode: 'View' | 'New' | 'Edit' | 'Attachments' | '';
   panelWidth: PanelType;
   showAttach: boolean;
   clickedAttach: boolean;  //if you clicked the attached icon (vs selected row), it only will show the attachments in the panel for cleaner implimentation
@@ -102,6 +129,10 @@ export interface IReactListItemsState {
 
   myDialog? : IMyDialogProps;
   pickedCommand?: IQuickButton; //Index of command and ID of panel item
+
+  allowSplit: boolean; //From Edit Panel
+  
+  recentUsers: IUser[];
 
 }
 
@@ -255,6 +286,11 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
             groupByFields.map( gF => {  gF.name = gF.name.replace('/','') ;  });
         }
 
+        //Create 
+        let recentUsers = JSON.parse(JSON.stringify(this.props.recentUsers ));
+
+        let quickFields : IQuickField[][] = JSON.parse(JSON.stringify( this.props.quickFields ));
+
         this.state = {
           maxChars: this.props.maxChars ? this.props.maxChars : 50,
           parentListFieldTitles:parentListFieldTitles,
@@ -271,6 +307,14 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
           myDialog: this.createBlankDialog(),
           pickedCommand: null,
           panelWidth: PanelType.medium,
+          allowSplit: this.props.allowSplit,
+          showNewPanel: false,
+          showEditPanel: false,
+          panelMode: '',
+          
+            recentUsers: recentUsers,
+            quickFields: quickFields,
+
         };
     }
         
@@ -296,6 +340,13 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
         if ( prevProps.viewFields !== this.props.viewFields ) { redraw = true; }
         if ( prevProps.items.length !== this.props.items.length ) { redraw = true; }
         if ( prevProps.parentListURL !== this.props.parentListURL ) { redraw = true; }
+
+        if ( prevProps.webURL !== this.props.webURL ) { redraw = true; }
+        if ( prevProps.listName !== this.props.listName ) { redraw = true; }
+
+
+        if ( prevProps.quickFields !== this.props.quickFields ) { redraw = true; }
+        if ( prevProps.quickCommands !== this.props.quickCommands ) { redraw = true; }
 
         //checking this does not redraw the component
 //        if ( prevProps.WebpartHeight !== this.props.WebpartHeight ) { redraw = true; }
@@ -325,7 +376,7 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
         let thisLog = null;
 
-        if ( this.props.items != null && this.props.items.length > 0 ) { 
+        if ( this.props.items != null ) { 
 
             let attachments = this.state.panelAttachments.length > 0 ? this.state.panelAttachments : null ;
 
@@ -346,21 +397,43 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
            //Only get buttons if panelItem is selected
 
             let buttons, fields, toggles = null;
+            let panelHeaderText = this.getPanelHeaderText();
 
             if ( this.state.showPanel === true ) {
+
                 buttons = createPanelButtons( this.props.quickCommands, this.state.panelItem, this._panelButtonClicked.bind(this), this.props.sourceUserInfo ) ;
                 toggles = <div style={{ float: 'right' }}> { makeToggles(this.getPageToggles( this.state.panelWidth )) } </div>;
-
+                
+                fields = this.state.panelMode !== 'New' ? null : 
+                <ThisEditPane 
+                    wpContext={ this.props.wpContext }
+                    webAbsoluteUrl={ this.props.webURL }
+                    fields = { this.state.quickFields }
+                    contextUserInfo = { this.props.contextUserInfo }
+                    sourceUserInfo = { this.props.sourceUserInfo }
+                    onChange = { this._editFieldUpdate.bind(this) }
+                    _clearDateField = { this._clearDateField.bind(this) }
+                    _addYouToField = { this._addUserToField.bind(this) }
+                    _addWeekToDate = { this._addWeekToDate.bind(this) }
+                    _updateDropdown = { this._updateDropdown.bind(this) }
+                    _saveItem= { this._saveItem.bind(this) }
+                    _cancelItem= { this._onClosePanel.bind(this) }
+                    allowSplit= { this.state.allowSplit }
+                    _getTitleValue = { null /*this.updatePageTitleInStateTest.bind(this)  null */  }
+                    readOnlyMode = { false }
+                ></ThisEditPane>;
             }
+
 
             let fullPanel = !this.state.showPanel ? null : 
                 <Panel
                     isOpen={this.state.showPanel}
                     type={ this.state.panelWidth }
                     onDismiss={this._onClosePanel}
-                    headerText={ this.state.panelId.toString() }
+                    headerText={ panelHeaderText }
                     closeButtonAriaLabel="Close"
                     onRenderFooterContent={this._onRenderFooterContent}
+                    onRenderHeader={ this.props.allowSplit ? this._onRenderHeader : null }
                     isLightDismiss={ true }
                     isFooterAtBottom={ true }
                 >
@@ -406,30 +479,46 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
 
             let viewFields = attachField.concat( viewFieldsBase );
 
-            let listView = <div>
-            <ListView
-                items={ this.props.items }
-                viewFields={ viewFields }
-                compact={true}
-                selectionMode={ this.props.includeDetails ? SelectionMode.single : SelectionMode.none }
-                selection={ this.props.updateParentComponentEditPane ? this.props.updateParentComponentEditPane : this._onShowPanel.bind(this) }
-                showFilter={false}
-                //defaultFilter="John"
-                filterPlaceHolder="Search..."
-                groupByFields={ this.props.groupByFields } 
-            /></div>;
+            let listView = null;
+            
+            if ( this.props.items.length === 0 ) {
+                listView = <div> There are no items to display yet</div>;
+
+            } else {
+                listView = <div>
+                <ListView
+                    items={ this.props.items }
+                    viewFields={ viewFields }
+                    compact={true}
+                    selectionMode={ this.props.includeDetails ? SelectionMode.single : SelectionMode.none }
+                    selection={ this._onShowPanelEditItem.bind(this) }
+                    showFilter={false}
+                    //defaultFilter="John"
+                    filterPlaceHolder="Search..."
+                    groupByFields={ this.props.groupByFields } 
+                /></div>;
+            }
+ 
 
             //        let logTable = itemRows === null ? <div>Nothing to show</div> : <table style={{ display: 'block'}} className={stylesInfo.infoTable}>
 
             let barText = this.props.blueBar && this.props.blueBar != null ? this.props.blueBar : <span>Items</span>;
 
+            let tipsStyles = defCommandIconStyles;
+            tipsStyles.root.padding = '2px';
+            tipsStyles.icon.fontSize = 24;
+            let toggleNewItemPane = createIconButton('CirclePlus','Create Item',this._onShowPanelNewItem.bind(this), null, tipsStyles );
+
             let webTitle = null;
+
             let listLink = !this.props.includeListLink ? null : <div className={ stylesInfo.infoHeading } onClick={ this._onGoToList.bind(this) } 
                 style={{ paddingRight: 20, whiteSpace: 'nowrap', float: 'right', paddingTop: 0, cursor: 'pointer', fontSize: 'smaller',background: 'transparent' }}>
-                    <span style={{ background: 'transparent' }} className={ stylesInfo.listLink }>Go to list</span></div>;
+                    <span style={{ background: 'transparent' }} className={ stylesInfo.listLink }>Go to list</span>
+                    <span style={{marginLeft: 20, float: 'right' }}> { toggleNewItemPane } </span>
+                    </div>;
 
             if ( barText != null ) {
-                webTitle =<div className={ [stylesInfo.infoHeading, stylesInfo.innerShadow].join(' ') }><span style={{ paddingLeft: 20, whiteSpace: 'nowrap' }}>( { this.props.items.length }  ) Items in: { barText }</span>{ listLink }</div>;
+                webTitle =<div className={ [stylesInfo.infoHeading, stylesInfo.innerShadow].join(' ') }><span style={{ paddingLeft: 20, whiteSpace: 'nowrap' }}>( { this.props.items.length }  ) { barText }</span>{ listLink }</div>;
 
             
             /*stylesL.reactListView*/
@@ -455,7 +544,9 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
                 );
             } 
 
-        } //if ( this.props.items != null && this.props.items.length > 0 ) {    
+        } else { //if ( this.props.items != null ) { 
+            return null;
+        }  
     } // Render
 
 
@@ -491,6 +582,16 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
     private _updateStateOnPropsChange(): void {
         this.setState({
         });
+    }
+    
+    private getPanelHeaderText () {
+        let panelHeaderText = null;
+
+        if ( this.state.panelMode === 'New') { panelHeaderText = 'Create New item(s)' ; }
+        else if ( this.state.panelMode === 'View') { panelHeaderText = 'Edit item' + this.state.panelId.toString() ; }
+        else if ( this.state.panelMode === 'Edit') { panelHeaderText = 'Edit item' + this.state.panelId.toString() ; }
+
+        return panelHeaderText;
     }
 
 /***
@@ -629,6 +730,272 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
     }
 
 
+
+
+
+
+
+
+
+
+/***
+ *    d8888b. d88888b  .d88b.  d8888b. db      d88888b      d8888b. d888888b  .o88b. db   dD d88888b d8888b. 
+ *    88  `8D 88'     .8P  Y8. 88  `8D 88      88'          88  `8D   `88'   d8P  Y8 88 ,8P' 88'     88  `8D 
+ *    88oodD' 88ooooo 88    88 88oodD' 88      88ooooo      88oodD'    88    8P      88,8P   88ooooo 88oobY' 
+ *    88~~~   88~~~~~ 88    88 88~~~   88      88~~~~~      88~~~      88    8b      88`8b   88~~~~~ 88`8b   
+ *    88      88.     `8b  d8' 88      88booo. 88.          88        .88.   Y8b  d8 88 `88. 88.     88 `88. 
+ *    88      Y88888P  `Y88P'  88      Y88888P Y88888P      88      Y888888P  `Y88P' YP   YD Y88888P 88   YD 
+ *                                                                                                           
+ *                                                                                                           
+ */
+
+private _addUserToField = (prop: string, valueX: any ): void => {
+    let e: any = event;
+    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
+    thisID = thisID.replace('EditFieldID','');
+
+    let quickFields = this.state.quickFields;
+
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.name === thisID ) { 
+          if (field.type.toLowerCase().indexOf('user') < 0 ) {
+            alert('Error in _addUserToField:  Trying to add user to non-user field!');
+          } else {
+            let value = field.value;
+            if ( value == null ) { value = []; }
+            value.push( this.props.contextUserInfo );
+            field.value = value;
+            this.updateRecentUsers( field.value, this.state.recentUsers, this.props.webURL );
+          }
+        }
+      });
+    });
+    //Then update the quickFields
+
+    this.setState({ quickFields: quickFields, });
+
+  }
+
+
+  /***
+ *    d8888b.  .d8b.  d888888b d88888b      d8888b. d888888b  .o88b. db   dD d88888b d8888b. 
+ *    88  `8D d8' `8b `~~88~~' 88'          88  `8D   `88'   d8P  Y8 88 ,8P' 88'     88  `8D 
+ *    88   88 88ooo88    88    88ooooo      88oodD'    88    8P      88,8P   88ooooo 88oobY' 
+ *    88   88 88~~~88    88    88~~~~~      88~~~      88    8b      88`8b   88~~~~~ 88`8b   
+ *    88  .8D 88   88    88    88.          88        .88.   Y8b  d8 88 `88. 88.     88 `88. 
+ *    Y8888D' YP   YP    YP    Y88888P      88      Y888888P  `Y88P' YP   YD Y88888P 88   YD 
+ *                                                                                           
+ *                                                                                           
+ */
+
+  private _addWeekToDate = (prop: string, value: any ): void => {
+
+    let e: any = event;
+    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
+    thisID = thisID.replace('EditFieldID','');
+
+    let quickFields = this.state.quickFields;
+
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.name === thisID ) { 
+
+          //Based on https://www.sitepoint.com/community/t/how-do-i-add-one-week-to-a-date/47817/2
+          let start = field.value ? field.value: new Date();
+
+          field.value = new Date( start.getTime() + msPerDay * value );
+        }
+      });
+    });
+    //Then update the quickFields
+    this.setState({ quickFields: quickFields, });
+  }
+
+  private _clearDateField = (prop: string, value: any ): void => {
+
+    let e: any = event;
+    let thisID = findParentElementPropLikeThis(e.target, 'id', 'EditFieldID', 15, 'begins');
+    thisID = thisID.replace('EditFieldID','');
+
+    let quickFields = this.state.quickFields;
+
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.name === thisID ) { field.value = null ;}
+      });
+    });
+    //Then update the quickFields
+    this.setState({ quickFields: quickFields, });
+
+  }
+
+  /***
+ *    d88888b d8888b. d888888b d888888b      db    db d8888b. d8888b.  .d8b.  d888888b d88888b 
+ *    88'     88  `8D   `88'   `~~88~~'      88    88 88  `8D 88  `8D d8' `8b `~~88~~' 88'     
+ *    88ooooo 88   88    88       88         88    88 88oodD' 88   88 88ooo88    88    88ooooo 
+ *    88~~~~~ 88   88    88       88         88    88 88~~~   88   88 88~~~88    88    88~~~~~ 
+ *    88.     88  .8D   .88.      88         88b  d88 88      88  .8D 88   88    88    88.     
+ *    Y88888P Y8888D' Y888888P    YP         ~Y8888P' 88      Y8888D' YP   YP    YP    Y88888P 
+ *                                                                                             
+ *                                                                                             
+ */
+
+  private _editFieldUpdate = ( prop: string, value: any ): void => {
+
+    let e: any = event;
+
+    let quickFields = this.state.quickFields;
+    
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.name === prop ) { 
+          field.value = value ;
+          console.log('found this item to update: ' , prop, value );
+
+          if ( field.type.toLowerCase().indexOf('user') === 0  ) {
+            this.updateRecentUsers( field.value, this.state.recentUsers, this.props.webURL );
+
+          } else if ( field.type.toLowerCase().indexOf('user') > 0  ) { //covers multiUser and splitUser
+            this.updateRecentUsers( field.value, this.state.recentUsers, this.props.webURL );
+
+          }
+        }
+      });
+    });
+    //Then update the quickFields
+
+    this.setState({ quickFields: quickFields, });
+
+  }
+
+  private async updateRecentUsers( theseUsers: IUser[], checkTheseUsers: IUser[] , webUrl: string ) {
+    let recentUsers = await ensureTheseUsers( theseUsers, checkTheseUsers, webUrl );
+    this.setState({
+      recentUsers: recentUsers,
+    });
+  }
+
+  private _updateDropdown = (prop: React.FormEvent<HTMLDivElement>, e , pickedOption ): void => {
+
+    let quickFields = this.state.quickFields;
+    let thisProp : any = prop;
+
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.name === thisProp ) { 
+          field.value = pickedOption.text ;
+          console.log('found this item to update: ' , thisProp, pickedOption.text );
+        }
+      });
+    });
+    //Then update the quickFields
+
+    // console.log('HERE IS Current QuickFields:', quickFields );
+    this.setState({ quickFields: quickFields, });
+  }
+
+/***
+ *    .d8888.  .d8b.  db    db d88888b      d888888b d888888b d88888b .88b  d88. 
+ *    88'  YP d8' `8b 88    88 88'            `88'   `~~88~~' 88'     88'YbdP`88 
+ *    `8bo.   88ooo88 Y8    8P 88ooooo         88       88    88ooooo 88  88  88 
+ *      `Y8b. 88~~~88 `8b  d8' 88~~~~~         88       88    88~~~~~ 88  88  88 
+ *    db   8D 88   88  `8bd8'  88.            .88.      88    88.     88  88  88 
+ *    `8888Y' YP   YP    YP    Y88888P      Y888888P    YP    Y88888P YP  YP  YP 
+ *                                                                               
+ *                                                                               
+ */
+
+  private async _saveItem ( ) {
+
+    let splitCount = 1;
+    let splitUsers = [];
+    let splitField = '';
+    //Get array of split users
+    let quickFields = this.state.quickFields;
+    
+    //Search through each row and field for name:
+    quickFields.map( fieldRow => {
+      fieldRow.map ( field => {
+        if ( field.type.toLowerCase().indexOf('split') > -1 ) { 
+          splitUsers = JSON.parse( JSON.stringify( field.value ));
+          splitField = field.name;
+          splitCount = field.value ? field.value.length : 0;
+        }
+      });
+    });
+
+    let results : any = null;
+    if ( splitCount === 1 ) {
+      let recentUsers = JSON.parse(JSON.stringify( this.state.recentUsers )); // Needed to prevent it from getting over-written in this function somewhere
+      results = await _saveEditPaneItem( this.props.webURL, this.props.listName, this.state.quickFields, this.props.staticFields, recentUsers );
+  
+    } else {
+
+      //Save each item individually - unless allowSplit !== true, then just set to first item in array
+      if ( splitCount > 1 && this.state.allowSplit !== true ) { splitCount = 1; }
+      for (let i = 0; i < splitCount; i++) {
+
+        quickFields.map( fieldRow => {
+          fieldRow.map ( field => {
+            if ( field.name === splitField ) { 
+              field.value = [ splitUsers[i] ];
+            }
+          });
+        });
+
+        let recentUsers = JSON.parse(JSON.stringify( this.state.recentUsers )); // Needed to prevent it from getting over-written in this function somewhere
+        results = await _saveEditPaneItem( this.props.webURL, this.props.listName, quickFields, this.props.staticFields, recentUsers );
+
+      }
+
+    }
+
+    let passed = results && results.data ? true : false;
+
+    if ( passed !== true ) {
+      //The save did not happend
+      console.log('was NOT ABLE TO SAVE ITEM');
+      
+      //Put back original splitUsers array - NOT needed if I clear this field after save.
+      quickFields.map( fieldRow => {
+        fieldRow.map ( field => {
+          if ( field.name === splitField ) { 
+            field.value = splitUsers;
+          }
+        });
+      });
+
+     alert('Your item was not saved...\nPossibly because you have a splitPersonField type with no value.');
+
+    } else {
+      alert('Your Action News item was just saved!');
+
+      this.props.reloadAllItems();
+
+    }
+    return null;
+
+  }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 /***
  *    .d8888. db   db  .d88b.  db   d8b   db      d8888b.  .d8b.  d8b   db d88888b db      
  *    88'  YP 88   88 .8P  Y8. 88   I8I   88      88  `8D d8' `8b 888o  88 88'     88      
@@ -639,6 +1006,30 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
  *                                                                                         
  *                                                                                         
  */
+
+
+
+    public _onShowPanelNewItem = ( item: any ): void => {
+        // public async _onShowPanelNewItem ( item: any ) {
+        //This sends back the correct pivot category which matches the category on the tile.
+        //sourceUserInfo
+        let quickFields : IQuickField[][] = JSON.parse(JSON.stringify( this.props.quickFields ));
+        this.setState({ quickFields: quickFields, showNewPanel: true, showEditPanel: false, showPanel: true, panelMode: 'New' });
+
+  
+    } //End toggleNewItem  
+
+    public _onShowPanelEditItem = ( item: any ): void => {
+        let quickFields : IQuickField[][] = JSON.parse(JSON.stringify( this.props.quickFields ));
+
+        quickFields.map( fieldRow => {
+            fieldRow.map( thisFieldObject => {
+                thisFieldObject.value = item[0][thisFieldObject.name];
+            });
+        }) ;
+
+        this.setState({ quickFields: quickFields, showNewPanel: false, showEditPanel: true, showPanel: true, panelMode: 'Edit' });
+    } //End toggleNewItem 
 
     private _onShowPanel = (item): void => {
   
@@ -678,6 +1069,8 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
                     showPanel: showFullPanel,
                     showAttach: showAttachPanel , 
                     clickedAttach: clickedAttach,
+                    showEditPanel: true,
+                    showNewPanel: false,
                     panelId: thisID,
                     panelItem: panelItem,
                     lastPanelId: lastPanelId,
@@ -705,6 +1098,8 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
     private _onClosePanel = (): void => {
         this.setState({ 
             showPanel: false,
+            showNewPanel: false,
+            showEditPanel: false,
             showAttach: false , 
             clickedAttach: false,
          });
@@ -740,7 +1135,37 @@ export default class ReactListItems extends React.Component<IReactListItemsProps
         );
     }
 
+    private _onRenderHeader = (): JSX.Element => {
+
+        let defStyles = { root: { width: 160, } };
+  
+        let thisToggle = <Toggle label={ 'Split Notifications' } 
+            onText={ 'On' } 
+            offText={ 'Off' } 
+            onChange={ this._toggleSplit.bind(this) } 
+            checked={ this.state.allowSplit }
+            styles={ defStyles }
+        />;
         
+        const stackTokens: IStackTokens = { childrenGap: 20 };
+        let headerText = this.getPanelHeaderText();
+        return (
+        <div>
+          <Stack horizontal={ true } horizontalAlign= { 'space-between' } tokens={stackTokens}>
+            <span style={{ marginLeft: 35, fontSize: 28, marginTop: 5 }}> { headerText } </span>
+            { thisToggle }
+          </Stack>
+  
+        </div>
+        );
+    }
+        
+    
+    private _toggleSplit() {
+
+        this.setState({ allowSplit: this.state.allowSplit !== true ? true : false });
+
+    }
     /***
      *         d888888b  .d88b.   d888b   d888b  db      d88888b .d8888. 
      *         `~~88~~' .8P  Y8. 88' Y8b 88' Y8b 88      88'     88'  YP
