+++ src/webparts/actionnews/components/const_ActionCommands.ts
@@ -5,7 +5,6 @@ export const ActionQuickCommands: IQuickCommands = {
     callBack: null,
     successBanner: 3.5,
 
-    quickFields: [],
     buttons: [
       [
         {
